Fix bug in services framework keys
PR #28 updated the framework keys for brief object stubs, but also
accidentally changed the framework keys for service object stubs. We
didn't notice at the time because we didn't update dmtestutils in the
api repo, but this change breaks the tests for the api app.

This commit changes the framework keys in service object stubs to be
back inline with the current model, and also tweaks the briefs
frameworks keys code to be more explicit.

Also, change `_format_framework`'s keyword arguments to be words so that
spellcheckers don't complain, and make them non-optional.