Test release and publish workflows
This commit should cause the new release and publish workflows to run
(see #35). Technically this is a downgrade of the version number (as rc1
is a [pre-release signifier][1]), but I intend to revert this PR once
I've tested the workflows work. Fingers crossed I don't need to do this
more than once!!

[1]: https://www.python.org/dev/peps/pep-0440/#pre-releases