Fix check if tag exists step in release workflow
Change the step `version_tag` to use an existing third-party action
instead of shell script.

It turns out that the the checkout action does not include tags (see the
output of workflow run [Create a GitHub release #2] which includes
debugging information), so the shell script used previously always
returns false. The third-party action instead uses the GitHub API to
check if the tag exists, which should always work.

[Create a GitHub release #2]: https://github.com/alphagov/digitalmarketplace-test-utils/runs/1630511744